requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
Babel==2.4.0
beautifulsoup4==4.5.3
billiard==3.3.0.23
bleach==1.4
bootstrap-admin==0.3.7.1
captcha==0.2.2
celery==3.1.25
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.10.6
django-admin-bootstrapped==2.5.7
django-admin-interface==0.3.2
django-allauth==0.31.0
django-authtools==1.5.0
django-autoslug==1.9.3
django-braces==1.11.0
django-celery==3.2.1
django-celery-results==1.0.1
django-chart-tools==1.0
django-chartit==0.2.9
django-chartjs==1.2
django-ckeditor==5.3.0
django-colorfield==0.1.12
django-countries==4.6.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-decouple==2.1
django-environ==0.4.2
django-extra-views==0.9.0
django-filter==1.0.4
django-flat-responsive==1.3.0
django-flat-theme==1.1.4
django-flatpages-x==0.1.5
django-formset-js==0.5.0
django-formtools==2.0
django-froala-editor==2.6.4
django-geoposition==0.3.0
django-haystack==2.6.0
django-import-export==0.5.1
django-jquery-js==3.1.1
django-js-asset==0.1.1
django-markdown==0.8.4
django-material==1.0.0
django-mathfilters==0.4.0
django-mathjax==0.0.8
django-modelcluster==3.1
django-mptt==0.8.7
django-pagedown==0.1.3
django-phonenumber-field==1.3.0
django-recaptcha==1.2.1
django-taggit==0.22.0
django-tinymce==2.6.0
django-todo==1.6.2
django-treebeard==4.1.0
django-user-sessions==1.5.0
django-uuidfield==0.5.0
django-widget-tweaks==1.4.1
django-wpadmin==1.7.4
django-youtube==0.2
djangorestframework==3.6.2
docker==2.2.1
docker-pycreds==0.2.1
docutils==0.13.1
easy-thumbnails==2.4.1
et-xmlfile==1.0.1
ez-setup==0.9
feedparser==5.1.3
gdata==2.0.18
gunicorn==19.7.1
html5lib==0.999
httplib2==0.9.2
jdcal==1.3
Jinja2==2.9.6
kombu==3.0.37
langdetect==1.0.5
Markdown==2.4.1
markdown2==2.3.3
MarkupSafe==0.23
material==0.1
oauthlib==0.6.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
ordereddict==1.1
packaging==16.8
phonenumberslite==8.5.1
Pillow==4.0.0
psycopg2==2.7.1
Pygments==1.6
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
pysolr==3.6.0
python-decouple==3.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
pyzmail==1.0.3
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.4.1
six==1.10.0
solr==0.4
sqlparse==0.2.3
tablib==0.11.5
unicodecsv==0.14.1
Unidecode==0.4.20
Unipath==1.1
urllib3==1.20
vine==1.1.3
webencodings==0.5.1
websocket-client==0.40.0
whitenoise==3.3.0
Whoosh==2.6.0
Willow==0.4
xhtml2pdf==0.2b1
xlrd==1.0.0
xlwt==1.2.0
